Added a PostgreSQL database and created calculator functions
This commit introduces substantial changes to `docker-compose.yml`. We've included a PostgreSQL database to the Docker container to use during local development. This service depends on the app service and we've mapped their mutual interdependencies. We've also created volumes to persist data beyond the lifetime of the container.

Along with these changes, we added new calculator functions in the `calc.py` file. This consists of simple addition and subtraction operations and tests for these operations are included in `tests.py`. This sets a foundation for more complex calculator operations and tests in future.